remove get-pot and move file to util
