+++ src/client/src/utils/constants.ts
@@ -104,7 +104,6 @@ const EXTENSION_COMMANDS = {
   TRACK_PAGE_SWITCH: "track-page-switch",
   GEN_STATUS_MESSAGE: "update-status-message",
   GEN_STATUS: "update-status",
-  GET_PORT: "get-port",
   OPEN_PROJECT_IN_VSCODE: "open-project-vscode",
   GET_VERSIONS: "get-versions",
   CLOSE_WIZARD: "close-wizard",
